requirements.txt
streamlit==1.14.1
streamlit-folium==0.11.1
plotly==5.6.0
pandas==1.4.2
numpy==1.21.5
folium==0.14.0
matplotlib==3.5.1
matplotlib-inline==0.1.2
haversine==2.8.0
pillow==9.0.1